feat: 테스트

+++ src/main/java/com/running/you_run/running/payload/dto/TrackRecordDto.java
@@ -3,6 +3,6 @@ package com.running.you_run.running.payload.dto;
 public record TrackRecordDto(
         Long userId,
         String username,
-        int duration
+        long duration
 ) {
 }
